Calculate person-months of effort internally
Migrate from querying the tool at http://softwarecost.org/tools/COCOMO/
to get person-months of effort to calculating it within the package.
The scale factor and cost driver values used represent the default of
"Nominal" that is used by the aforementioned tool. The value for
labor_hours is rounded to the tenths for consistency with output from
the previous version.