Separate stdout and stderr in the `execute()` helper function
If stderr is piped into stdout analyzing a repository will fail if any
file fails during cloc's processing. This is most problematic for
projects that contain minified JavaScript code. The design of cloc is
around typical multi-line source files. Large minified JavaScript will
run the risk of causing analysis to fail with a "Line count, exceeded
timeout" error. This then causes the entire cloc analysis to fail due
to the design of the execute() function. With this change to execute()
I added a warning if stderr is not empty following the cloc call.